Split out timeouts.rs and refactor timeout calculation
